update login and booking

+++ src/main/java/entity/User.java
@@ -3,13 +3,13 @@ package entity;
 import java.io.Serializable;
 
 public class User implements Serializable, Identifiable {
-    final int id;
+    final long id;
     private final String login;
     private final String password;
 
     private static final long serialVersionUID = 4L;
 
-    public User(int id, String login, String password) {
+    public User(long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
